pyproject.toml
[tool.poetry]
name = "hokusai"
version = "0.5.10"
description = "Hokusai is a Docker + Kubernetes CLI for application developers."
authors = ["Isac Petruzzi <isac.petruzzi@gmail.com>"]

[tool.poetry.dependencies]
boto3 = "==1.9.162"
botocore = "==1.12.162"
click = "==6.7"
click-repl = "==0.1.6"
termcolor = "==1.1.0"
MarkupSafe = "==1.1.1"
Jinja2 = "==2.10.1"
prompt_toolkit = "==1.0.16"
PyYAML = "==3.13"

[tool.poetry.dev-dependencies]
pyinstaller = "==3.3.1"
httpretty = "==0.8.14"
mock = "==2.0.0"
packaging = "~18.0"
twine = "*"
wheel = "*"
ipython = "*"
ipdb = "*"
jupyter = "*"
nbformat = "==4.4.0"
pylint = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
